test_suite/testcase/saasWeb/__init__.py
# coding:utf-8

import logging

import allure
import pytest

from common.checkResult import asserting
from util.tools.log import Log
from util.tools.readYamlFile import ini_allyaml
from common.basePage import apisend
from util.tools.iniRequests import relevance
alldata = ini_allyaml()

Log()
__all__ = [
    'pytest',
    'asserting',
    'Log',
    'alldata',
    'logging',
    'allure',
    'apisend',
    'alldata',
    'relevance',
]