requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
blinker==1.4
Brotli==1.0.9
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
coverage==5.5
cryptography==3.4.8
DingtalkChatbot==1.5.3
Flask==2.0.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonpath==0.82
kaitaistruct==0.9
ldap3==2.9.1
MarkupSafe==2.0.1
mitmproxy==7.0.4
msgpack==1.0.3
packaging==20.9
passlib==1.7.4
pluggy==0.13.1
protobuf==3.18.1
publicsuffix2==2.20191221
py==1.10.0
pyasn1==0.4.8
pycparser==2.21
pydivert==2.1.0
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pytest==6.2.4
pytest-ordering==0.6
pytest-rerunfailures==10.0
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.2
six==1.16.0
sortedcontainers==2.4.0
toml==0.10.2
tornado==6.1
urllib3==1.26.5
urwid==2.1.2
Werkzeug==2.0.2
wsproto==1.0.0
zstandard==0.15.2
faker==13.3.2
pytest-xdist==2.5.0
